+++ src/main/java/com/nhnacademy/minidooray/task/backend/service/TaskServiceImpl.java
@@ -1,6 +1,7 @@
 package com.nhnacademy.minidooray.task.backend.service;
 
 import com.nhnacademy.minidooray.task.backend.domain.CommentDto;
+import com.nhnacademy.minidooray.task.backend.domain.CommentModifyRequest;
 import com.nhnacademy.minidooray.task.backend.domain.CommentRequest;
 import com.nhnacademy.minidooray.task.backend.domain.TaskDto;
 import com.nhnacademy.minidooray.task.backend.domain.TaskRequest;
@@ -13,7 +14,7 @@ import com.nhnacademy.minidooray.task.backend.repository.MilestoneRepository;
 import com.nhnacademy.minidooray.task.backend.repository.ProjectRepository;
 import com.nhnacademy.minidooray.task.backend.repository.TaskRepository;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Optional;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -41,22 +42,29 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
-    public TaskDto findTask(Long projectId, Long taskId) {
-        return taskRepository.findTaskByProjectIdAndId(projectId, taskId);
+    public Optional<TaskDto> findTask(Long taskId) {
+        return taskRepository.findTaskById(taskId);
     }
 
     @Override
-    public void createTask(TaskRequest taskRequest, Long projectId) {
-        Project project = projectRepository.getProjectById(projectId);
+    public boolean createTask(TaskRequest taskRequest) {
+        Project project = projectRepository.getProjectById(taskRequest.getProjectId());
         Milestone milestone = milestoneRepository.findById(taskRequest.getMilestoneId()).orElse(null);
-        Task task = new Task(taskRequest.getName(), project, milestone);
+        Task task = Task.builder().name(taskRequest.getName()).project(project).milestone(milestone).build();
 
-        taskRepository.saveAndFlush(task);
+        taskRepository.save(task);
+        return true;
     }
 
     @Override
-    public void deleteTask(Long taskId) {
-        taskRepository.deleteTaskById(taskId);
+    public boolean deleteTask(Long taskId) {
+        if (taskRepository.existsById(taskId)) {
+            taskRepository.deleteTaskById(taskId);
+            return true;
+        } else {
+            return false;
+        }
+
     }
 
     @Override
@@ -65,15 +73,46 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
-    public void createComment(CommentRequest commentRequest, Long taskId) {
+    public boolean createComment(CommentRequest commentRequest, Long taskId) {
         Task task = taskRepository.getTaskById(taskId);
-        Comment comment = new Comment(commentRequest.getContent(), task);
+        Comment comment = Comment.builder().content(commentRequest.getContent()).task(task).build();
 
-        commentRepository.saveAndFlush(comment);
+        commentRepository.save(comment);
+
+        return true;
     }
 
     @Override
-    public void deleteComment(Long taskId) {
-        commentRepository.deleteById(taskId);
+    public boolean modifyComment(Long commentId, CommentModifyRequest commentModifyRequest) {
+        Optional<Comment> commentOptional = commentRepository.findById(commentId);
+        if (commentOptional.isPresent()) {
+            Comment comment = commentOptional.get();
+            comment = Comment.builder().id(commentModifyRequest.getId()).content(commentModifyRequest.getContent())
+                    .build();
+            commentRepository.save(comment);
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    @Override
+    public boolean modifyTask(Long taskId, TaskRequest taskRequest) {
+        Project project = projectRepository.getProjectById(taskRequest.getProjectId());
+        Task task = Task.builder().name(taskRequest.getName()).project(project).build();
+        taskRepository.save(task);
+        return true;
+    }
+
+    @Override
+    public boolean deleteComment(Long taskId) {
+        if (taskRepository.existsById(taskId)) {
+            commentRepository.deleteById(taskId);
+            return true;
+
+        } else {
+            return false;
+        }
+
     }
 }
